requirements.txt
git+https://github.com/nghorbani/configer@5f2b82e5
git+https://github.com/nghorbani/human_body_prior@1936f38
git+https://github.com/nghorbani/amass@e5f0889
--find-links https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torch==1.8.2 #+cu111
pytorch-ignite
scipy
PyYAML==5.3.1
numpy
tqdm
tensorboardX
zarr
pandas
matplotlib
trimesh